==> src/1.1-routing-engine-preparation.ipynb <==
## Prepare OSM data and convert it into routable graph

%load_ext autoreload
%autoreload 2
%cd C:\emobility
%matplotlib inline

# Load libs
import pandas as pd
import os
import geopandas as gpd
import dataworkers as dw
from shapely.geometry import box
import requests as req
from pprint import pprint
from lib import routing

### 1. Prepare .poly file for cropping OSM data

# Compute the boundary of VG car agents' activities
df_act = pd.read_csv('dbs/agents/1_activity_plans.csv')
df_act.head()

gdf_act = dw.df2gdf_point(df_act, 'POINT_X', 'POINT_Y', crs=3006, drop=True).to_crs(4326)
geom = box(*gdf_act.total_bounds)
gdf_box = gpd.GeoDataFrame([1], geometry=[geom], crs=4326)
gdf_box = gdf_box.rename(columns={0: 'box'})
gdf_box

dw.gdf2poly(gdf=gdf_box, targetfile='dbs/vg_car_agents_act_bd.poly')

### 2. Use osmosis to crop the data
osmosis \--read-pbf file="sweden-latest.osm.pbf" \--bounding-polygon file="vg_car_agents_act_bd.poly" \--write-pbf file="sweden_vg_car_users.osm.pbf"

## 3. Try OTP with elevation as the routing engine

# 4 Build routable OTP graph
region = 'vg'
otp_file = 'dbs/otp-2.2.0-shaded.jar'
otp_folder =  f'dbs/{region}'
memory_gb = 63 # assigned memory for java in GB
if os.path.exists(otp_folder + '/Graph.obj') is not True:
    print(f"{region}: build routable OTP graph.")
    routing.otp_build(otp_file=otp_file, otp_folder=otp_folder, memory_gb=memory_gb)
else:
    print(f"Graph.obj for {region} exists. Graph building skipped.")

Start playing with OTP API from here below.

country = 'sweden'
region = 'vg'
# Start the server
print('Start the OTP server...')
otp_file = 'dbs/otp-2.2.0-shaded.jar'
otp_folder = f'dbs/{region}'
memory_gb = 40  # assigned memory for java in GB
routing.otp_server_starter(otp_file=otp_file, otp_folder=otp_folder, memory_gb=memory_gb)

fromPlace = (59.40223829, 18.07072809)
toPlace = (59.33297, 18.04428)
maxWalkDistance = 300
mode = "BICYCLE,WALK"


request = routing.req_define(fromPlace=fromPlace,
                     toPlace=toPlace,
                     maxWalkDistance=maxWalkDistance,
                     time="12:00am",
                     date="3-24-2023",
                     mode=mode)

request

resp = req.get(request)
output = resp.json()

plan = output["plan"]["itineraries"][0]
pprint(plan)
