==> src/1.2-activity-plan-preparation.ipynb <==
import pandas as pd
import folium
import geopandas as gpd
import numpy as np
import warnings
warnings.filterwarnings('ignore')

%cd C:\emobility

### Convert SySMo's activity_plans to OTP input data format

df_act = pd.read_csv(f'./dbs/agents/1_activity_plans.csv')


# coordinate system conversion

df_act= gpd.GeoDataFrame(df_act, geometry=gpd.points_from_xy(x=df_act.POINT_X, y=df_act.POINT_Y), crs = 'EPSG:3006')
df_act = df_act.to_crs(4326)

df_act['lat']=df_act['geometry'].apply(lambda p: p.y)
df_act['lng']=df_act['geometry'].apply(lambda p: p.x)

max_trip_number = df_act.groupby(["person"])["act_id"].max().reset_index()
max_trip_number.rename(columns = {"act_id": "max_trip_number"}, inplace=True)

#do pivot table by activity ID's

df_act= df_act[["person", "act_id", "deso", "act_purpose", "lat", "lng"]]

df_act = df_act.pivot_table(index='person', columns= 'act_id', values=[ "lat", "lng", "deso", "act_purpose"], aggfunc = lambda x: ','.join(str(v) for v in x) )

df_act= pd.merge(df_act, max_trip_number, left_index=True, right_on='person', how='left')
#df_act.columns

df_OTP = pd.DataFrame(columns=['person', 'act_id', 'act_purpose', 'lat_o', 'lng_o', 'lat_d', 'lng_d', 'origin', 'destination'])

for max_trip in range(2, max(max_trip_number.max_trip_number)+1):
    df_something = df_act[df_act.max_trip_number==max_trip]

    for trip in range(max_trip):
        df_something['act_id']=trip
        df_OTP = pd.DataFrame(np.vstack([df_OTP.values, df_something[['person', 'act_id', ('act_purpose',trip+1), ('lat',trip), ('lng',trip), ('lat',trip+1), ('lng',trip+1), ('deso',trip), ('deso',trip+1)]].values]), columns=df_OTP.columns)


df_OTP = df_OTP.rename_axis('MyIdx').sort_values(by = ['person', 'MyIdx'], ascending = [True, True])
df_OTP['origin'] = df_OTP['origin'].str[:4].astype(int)
df_OTP['destination'] = df_OTP['destination'].str[:4].astype(int)
df_OTP[['lat_o', 'lng_o', 'lat_d', 'lng_d']] = df_OTP[['lat_o', 'lng_o', 'lat_d', 'lng_d']].astype(float)

#df_OTP.head()

### calculate an average cycling speed by age,gender and trip purposes for each person

from avg_speed_calculator import main


if __name__ == '__main__':
    df_OTP = main(df_OTP)

df_OTP.tail()

### divede the data for multiprocess routing and save

# divide the data to 30 parts
splits = np.array_split(df_OTP.person, 30)
df_OTP['part'] = 0


count=0
for split in range(30):
    df_OTP['part'][df_OTP.person.isin(splits[split])]=count
    count+=1
    print(count)

# save the data
df_OTP.to_csv(f'./dbs/agents/od_pairs_vg.csv', index=False)
