==> src/4.1-emission_model.ipynb <==
import numpy as np
import pandas as pd
import os, sys
import warnings
from tqdm import tqdm

warnings.filterwarnings('ignore')


%cd C:\emobility

## Car assignment

df_act = pd.read_csv(f'./dbs/agents/1_activity_plans.csv')


emission_factor = pd.read_csv(f'./dbs/emission/EmissionFactor.csv', delimiter=";")
PHEV = emission_factor[(emission_factor['carType']== 'Car_PHEV_bensin') | (emission_factor['carType']== 'Car_PHEV_diesel')].mean(axis=0)
emission_factor = emission_factor.append( PHEV, ignore_index=True ).drop([4, 5], axis=0)
emission_factor.iloc[5,0] = 'Car_PHEV'
emission_factor.replace(['Car_gasoline','Car_diesel','Car_ethanol','Car_gas','Car_el ','Car_PHEV'],[0,1,4,5,2,3], inplace=True)




car_distribution = pd.read_csv(f'./dbs/emission/CarDistribution.csv', delimiter=";")
car_distribution.loc[:,['Gasoline', 'Diesel', 'El', 'PHEV', 'Ethanol', 'Gas']] = car_distribution.loc[:,['Gasoline', 'Diesel', 'El', 'PHEV', 'Ethanol', 'Gas']].div(car_distribution['Total'], axis=0)

car_distribution = car_distribution.drop(['Total'], axis=1)



# generate individual table with total travel dist column
df_act_car_segments = df_act.drop_duplicates(subset=['person'])

df_act_car_segments['carType'] = 100

df_act_car_segments_total_dist = df_act.groupby(['person'])['distance'].sum().reset_index(name='total_dist')
df_act_car_segments = pd.merge(df_act_car_segments, df_act_car_segments_total_dist,how='left',on='person')

df_act_car_segments_total_dist.total_dist.median()

## Adjusting Diesel Probability Distributions Based on total distance travelled

To adjust $p_1$ based on 'total_dist', a function can be applied to increase or decrease $p_1$ in correlation with 'total_dist', while ensuring it remains within the probability limits of 0 and 1. Adjustments to $p_1$ require corresponding adjustments to $p_2$ and $p_3$ to ensure the total probability remains at 1:
- If $p_1$ is increased, $p_2$ and $p_3$ should decrease proportionally or other way around.

To modify the first element of the probability distribution:
- $\[ p_1' = p_1 + Adjustment Sensitivity (k) * (total Dist - total Dist Median) \] $

Adjusting the Remaining Probabilities:

- $\[ p_2' = p_2 + \frac{1}{2}(p_1 - p_1') \]$

- $\[ p_3' = p_3 + \frac{1}{2}(p_1 - p_1') \]$

### Calculating the Adjustment Sensitivity (k)

The responsiveness of $p_1$ to changes in $a$ is controlled by a scaling factor $k$. To determine the appropriate value for $k$, consider the range over which 'totalDist' varies:
- Set boundaries for $p_1'$ (the adjusted $p_1$) to ensure it remains within the open interval.
- Here, we set the max value for $p_1$ to 0.75.
This can be calculated using the formulas:

$k_{\text{max}}$ = $\frac{0.75 - p_1}{totalDistMax - totalDistMedian}$


# Compute k values showing the sensitivity to the probability adjustment for each deso
def compute_k_values(df, car_distribution):
    deso_k_values = {}
    total_dist_max = df['total_dist'].max()
    total_dist_median = df['total_dist'].median()

    for deso in df['deso'].unique():
        if deso[:2] == '14':
            gasoline_prob = car_distribution[['Diesel']][car_distribution['Kod'] == int(deso[:4])].values[0][0]
            k_max = (0.75 - gasoline_prob) / (total_dist_max - total_dist_median) if (total_dist_max - total_dist_median) != 0 else 0
            deso_k_values[deso] = k_max
    return deso_k_values


# Adjust the Gasoline probability based on total_dist and k values for each deso
def adjust_probability(row, prob, k_values):
    k = k_values.get(row['deso'], 0)
    adjusted_prob = prob[0].copy()
    # print(adjusted_prob)
    adjusted_prob[1] += k * (row['total_dist'] - 33.62704)

    adjusted_prob[1] = max(min(adjusted_prob[1], 0.99), 0.01)
    adjusted_prob /= np.sum(adjusted_prob)
    # print(adjusted_prob)

    return adjusted_prob


# car assignment algorithm
def carType_assignment(row, car_distribution, k_values):
    if row['deso'][:2] == '14':
        prob = car_distribution[['Gasoline', 'Diesel', 'El', 'PHEV', 'Ethanol', 'Gas']][car_distribution.Kod==int(row['deso'][:4])].values.tolist()
        adjusted_prob = adjust_probability(row, prob, k_values)
        pred = np.random.multinomial(1, adjusted_prob, 1)
        carType = pred[0].tolist().index(1)
    else:
        # for those not in VG, use VG general distribution
        prob = [0.579957, 0.354470, 0.003693, 0.006844, 0.043670, 0.011365]
        pred = np.random.multinomial(1, prob, 1)
        carType = pred[0].tolist().index(1)

    return carType



# Compute k values
k_values = compute_k_values(df_act_car_segments, car_distribution)

# assign a car type to each ind
df_act_car_segments['carType'] = df_act_car_segments.apply( lambda row: carType_assignment(row, car_distribution, k_values), axis=1)

print('car distribution by fuel type at the aggregated level')
df_act_car_segments.carType.value_counts(normalize=True)

df_act_car_segments['total_dist'][df_act_car_segments['carType']== 0].mean()


df_act_car_segments['total_dist'][df_act_car_segments['carType']== 1].mean()

## Emission calculation
### the road network classification by the population density in DeSO where the road link passes

import geopandas as gpd

#A DeSO in category A is rural areas. B is mostly located in a suburban areas. C is mostly located in the municipality's central location.
# read the road network and DeSO zones

road_network = gpd.read_file(f'./dbs/emission/road/5_road_network_with_slope.shp')

road_network = road_network[['link_id','geometry']]
road_network = road_network.reset_index()

DeSO = gpd.read_file(f'./dbs/emission/deso_statistik_shp/Fordon_region.shp')
DeSO = DeSO[['Deso','geometry']]
DeSO['deso_category'] = DeSO['Deso'].str[4]


DeSO = DeSO.to_crs(road_network.crs)

from spatial_operations import spatial_join


if __name__ == '__main__':
    road_network = spatial_join(road_network, DeSO)

# the links crossing multiple DeSO are duplicated in the spatial join process.
# the link id is no unique, and it crosses multiple DeSO zones when it be combined

road_network_Noduplicate = road_network.copy()

road_network_Noduplicate['deso_category'] = road_network_Noduplicate['deso_category'].fillna('M')


road_network_Noduplicate = road_network_Noduplicate.groupby(['link_id']).agg({'deso_category': ' '.join}).reset_index()

#road_network_Noduplicate.deso_category.value_counts()

def deso_categories(row):
    if len(row['deso_category'])==1:
        deso_category = row['deso_category']
    elif 'M' in row['deso_category']:
        deso_category = 'M'
    else:
        deso_category = "".join(set(row['deso_category'])).replace(" ", "")
        if (len(deso_category)>1) & ("C" in deso_category) :
            deso_category = 'M'
        elif (len(deso_category)>1) :
            deso_category = 'B'
        #elif: (len(deso_category)>1) & ("A" in deso_category)  :
            #deso_category = 'B'

    return deso_category

road_network_Noduplicate['deso_category'] = road_network_Noduplicate.apply( lambda row: deso_categories(row), axis=1)


road_network_Noduplicate['deso_category'].value_counts(normalize=True)

road_network_Deso = road_network.copy()
road_network_Deso['Deso'] = road_network_Deso['Deso'].fillna('M')
road_network_Deso = road_network_Deso.groupby('link_id')['Deso'].apply(list).reset_index()


### car movement trajectories on the road network

df_act_car_tra = pd.read_csv(f'./dbs/emission/Car_trajectories.csv', delimiter=",")


df_act_car_tra.sort_values(["person", "seq"], inplace=True)

df_act_car_tra = pd.merge(df_act_car_tra[['person', 'seq', 'time', 'purpose', 'type', 'link', 'distance_driven']], road_network_Noduplicate[['link_id', 'deso_category']], left_on= "link", right_on="link_id", how="left")

df_act_car_tra = pd.merge(df_act_car_tra, road_network_Deso, left_on= "link", right_on="link_id", how="left")

df_act_car_tra['distance_driven_shift'] = df_act_car_tra['distance_driven'].shift(-1)
df_act_car_tra['person_shift'] = df_act_car_tra['person'].shift(-1).fillna(0).astype('int')

df_act_car_tra['link_distance']=0

df_act_car_tra['link_distance'] = df_act_car_tra.apply( lambda row: row['distance_driven_shift']-row['distance_driven'] if row['person'] == row['person_shift'] else 0, axis=1)

# According to link distance (I have checked in Arcmap), shift rows.
df_act_car_tra['link_distance'] = df_act_car_tra['link_distance'].shift(1).fillna(0)

### activity purpose deduction

df_OTP = pd.read_csv(f'./dbs/agents/od_pairs_vg.csv')

df_OTP_last_activity = df_OTP.drop_duplicates(subset=['person'], keep='last')


df_act_last_activity = df_act_car_tra[df_act_car_tra.type == 'vehicle leaves traffic']

df_act_last_activity = df_act_last_activity.drop_duplicates(subset=['person','type'], keep='last')

df_act_last_activity = pd.merge(df_act_last_activity[['person','seq']], df_OTP_last_activity[["person", "act_purpose"]], left_on= "person", right_on="person", how="left")


df_act_car_tra = pd.merge(df_act_car_tra, df_act_last_activity, on= ["person",'seq'], how="left")

df_act_car_tra['purpose'] = np.where(pd.isna(df_act_car_tra.purpose), df_act_car_tra['act_purpose'], df_act_car_tra['purpose'])

df_act_car_tra.sort_values(["person", "seq"], inplace=True)

df_act_car_tra['purpose'].bfill(axis ='rows', inplace=True)


df_act_car_tra.drop(['person_shift', 'distance_driven_shift','act_purpose', 'link_id_x', 'link_id_y'], inplace=True, axis=1)

### assign activity IDs

act_id = df_act_car_tra[df_act_car_tra.type == 'vehicle leaves traffic']
act_id['act_id'] = act_id.groupby('person').cumcount()

df_act_car_tra = pd.merge(df_act_car_tra, act_id[['act_id']], left_index=True, right_index=True, how="left")



df_act_car_tra['act_id'].bfill(axis ='rows', inplace=True)
df_act_car_tra['act_id'] = df_act_car_tra['act_id'].astype('int')



### calculate car emission by the links

### add car with the fuel type to the trajectories

df_act_car_tra = pd.merge(df_act_car_tra, df_act_car_segments[["person", "carType"]],  left_on="person", right_on="person", how="left")

df_act_car_tra = pd.merge(df_act_car_tra, emission_factor, left_on="carType",
                                       right_on="carType", how="left")

def link_emission(row):
    if row['deso_category'] == 'C':
        link_emission = row['link_distance'] * row['TTW_city']
    elif row['deso_category'] == 'M':
        link_emission = row['link_distance'] * row['TTW_mix']
    elif (row['deso_category'] == 'A') | (row['deso_category'] == 'B'):
        link_emission = row['link_distance'] * row['TTW_rural']
    return link_emission




df_act_car_tra['link_emission'] = df_act_car_tra.apply(lambda row: link_emission(row), axis=1)
df_act_car_tra['link_emission'].sum()

df_act_car_tra.to_pickle(f"./dbs/output/emission_bylink.pkl")





# df_act_car_tra_bycategories = df_act_car_tra.groupby(['person', 'act_id', 'deso_category'])[
#     'link_distance'].sum().reset_index()
#
# ### add car with the fuel type to the trajectories
#
# df_act_car_tra_bycategories = pd.merge(df_act_car_tra_bycategories, df_act_car_segments[["person", "carType"]], left_on="person", right_on="person", how="left")
# df_act_car_tra_bycategories = pd.merge(df_act_car_tra_bycategories, emission_factor, left_on="carType", right_on="carType", how="left")

# df_act_car_tra_bycategories = df_act_car_tra.groupby(['person','act_id','deso_category'])['link_distance'].sum().reset_index()
#

# ### add car with the fuel type to the trajectories
#
# df_act_car_tra_bycategories = pd.merge(df_act_car_tra_bycategories, df_act_car_segments[["person", "carType"]], left_on= "person", right_on="person", how="left")

# df_act_car_tra_bycategories = pd.merge(df_act_car_tra_bycategories, emission_factor, left_on= "carType", right_on="carType", how="left")

# def link_emission(row):
#     if row['deso_category']=='C':
#         link_emission = row['link_distance']*row['TTW_city']
#     elif row['deso_category']=='M':
#         link_emission = row['link_distance']*row['TTW_mix']
#     elif (row['deso_category']=='A')|(row['deso_category']=='B'):
#         link_emission = row['link_distance']*row['TTW_rural']
#     return link_emission

# df_act_car_tra_bycategories['link_emission'] = df_act_car_tra_bycategories.apply( lambda row: link_emission(row), axis=1)

# df_act_car_tra_bycategories['link_emission'].sum()

# df_act_car_byactivities = df_act_car_tra_bycategories.groupby(['person','act_id'])['link_emission'].sum().reset_index()
#
